[refactor] reorder components
