+++ src/Components/CardTableContainer/TableToolBar/CampaignToolBar/ButtonContained/ButtonContained.tsx
@@ -1,0 +1,14 @@
+import Button from '@mui/material/Button';
+import AddIcon from '@mui/icons-material/Add';
+
+export const ButtonContained = () => {
+    return (
+        <>     
+                <Button variant="contained">
+                    <AddIcon />
+                    CREAR CAMPAÑA
+                </Button>
+         
+        </>
+    )
+}
