+++ src/Components/ClippedDrawer/ClippedDrawer.tsx
@@ -1,15 +1,11 @@
-
-
 import CssBaseline from '@mui/material/CssBaseline';
 import Box from '@mui/material/Box';
 import Toolbar from '@mui/material/Toolbar';
-import Header from '../Header/Header';
-import { ListItems } from '../ListItems/ListItems';
+import Header from './Header/Header';
+import { SideBar } from './SideBar/SideBar';
 
-import BasicCard from '../CardContainer/Card';
-import { CampaignToolBar } from '../CardContainer/CampaignToolBar/CampaignToolBar';
-
-
+import { CardTableContainer } from '../CardTableContainer/CardTableContainer';
+import { CampaignToolBar } from '../CardTableContainer/TableToolBar/CampaignToolBar/CampaignToolBar';
 
 export const ClippedDrawer = () => {
     return (
@@ -17,11 +13,11 @@ export const ClippedDrawer = () => {
             <Box sx={{ display: 'flex' }}>
                 <CssBaseline />
                 <Header />
-                <ListItems />
+                <SideBar />
                 <Box component="main" sx={{ flexGrow: 1, p: 3 }} style={{ background: '#FAFBFF' }}>
-                    <Toolbar />           
-                    <CampaignToolBar/>
-                    <BasicCard />
+                    <Toolbar />
+                    <CampaignToolBar />
+                    <CardTableContainer />
                 </Box>
             </Box>
         </>
